+++ Application/Sources/Models/User.swift
@@ -11,13 +11,4 @@ import CoreData
 
 extension User: Identifiable {
 	typealias RawIdentifier = Int64
-
-	static var current: User? {
-		return current(in: DB.shared.view)
-	}
-
-	static func current(in moc: NSManagedObjectContext) -> User? {
-		guard let userID = Settings.shared.currentUserID else { return nil }
-		return try? moc.first(value: userID) as? User
-	}
 }